+++ R/get_contrast_data_factor.R
@@ -1,8 +1,8 @@
 get_contrast_data_factor <- function(model,
                                      newdata,
                                      variable,
-                                     interaction,
-                                     first_interaction,
+                                     cross,
+                                     first_cross,
                                      ...) {
 
     data.table::setDT(newdata)
@@ -10,12 +10,10 @@ get_contrast_data_factor <- function(model,
     if (is.factor(newdata[[variable$name]])) {
         levs <- levels(newdata[[variable$name]])
         convert_to_factor <- TRUE
+
     } else {
 
-        msg <- format_msg(
-        "The `%s` variable is treated as a categorical (factor) variable, but the
-        original data is of class %s. It is safer and faster to convert such variables
-        to factor before fitting the model and calling `marginaleffects` functions.")
+        msg <- "The `%s` variable is treated as a categorical (factor) variable, but the original data is of class %s. It is safer and faster to convert such variables to factor before fitting the model and calling `marginaleffects` functions." 
         msg <- sprintf(msg, variable$name, class(newdata[[variable$name]])[1])
         warn_once(msg, "marginaleffects_warning_factor_on_the_fly_conversion")
         original_data <- hush(insight::get_data(model))
@@ -29,7 +27,7 @@ get_contrast_data_factor <- function(model,
     }
 
     # index contrast orders based on variable$value
-    if (variable$value == "reference") {
+    if (isTRUE(variable$value == "reference")) {
         # null contrasts are interesting with interactions
         if (!isTRUE(interaction)) {
             levs_idx <- data.table::data.table(lo = levs[1], hi = levs[2:length(levs)])
@@ -37,7 +35,7 @@ get_contrast_data_factor <- function(model,
             levs_idx <- data.table::data.table(lo = levs[1], hi = levs)
         }
 
-    } else if (variable$value == "pairwise") {
+    } else if (isTRUE(variable$value == "pairwise")) {
         levs_idx <- CJ(lo = levs, hi = levs, sorted = FALSE)
         # null contrasts are interesting with interactions
         if (!isTRUE(interaction)) {
@@ -45,19 +43,40 @@ get_contrast_data_factor <- function(model,
             levs_idx <- levs_idx[match(levs_idx$lo, levs) < match(levs_idx$hi, levs),]
         }
 
-    } else if (variable$value == "all") {
+    } else if (isTRUE(variable$value == "all")) {
         levs_idx <- CJ(lo = levs, hi = levs, sorted = FALSE)
 
-    } else if (variable$value == "sequential") {
+    } else if (isTRUE(variable$value == "sequential")) {
         levs_idx <- data.table::data.table(lo = levs[1:(length(levs) - 1)], hi = levs[2:length(levs)])
+
+    } else if (length(variable$value) == 2) {
+        if (is.character(variable$value)) {
+            tmp <- newdata[[variable$name]]
+            if (any(!variable$value %in% as.character(tmp))) {
+                msg <- "Some of the values supplied to the `variables` argument were not found in the dataset."
+                insight::format_error(msg)
+            }
+            idx <- match(variable$value, as.character(tmp))
+            levs_idx <- data.table::data.table(lo = tmp[idx[1]], hi = tmp[idx[[2]]])
+        } else if (is.numeric(variable$value)) {
+            tmp <- newdata[[variable$name]]
+            levs_idx <- data.table::data.table(
+                lo = factor(as.character(variable$value[1]), levels = levels(tmp)),
+                hi = factor(as.character(variable$value[2]), levels = levels(tmp)))
+        } else {
+            levs_idx <- data.table::data.table(lo = variable$value[1], hi = variable$value[2])
+        }
     }
 
     # internal option applied to the first of several contrasts when
     # interaction=TRUE to avoid duplication. when only the first contrast
     # flips, we get a negative sign, but if first increases and second
     # decreases, we get different total effects.
-    if (isTRUE(first_interaction)) {
-        levs_idx <- levs_idx[match(levs_idx$hi, levs) >= match(levs_idx$lo, levs),]
+    if (isTRUE(first_cross)) {
+        idx <- match(levs_idx$hi, levs) >= match(levs_idx$lo, levs)
+        if (sum(idx) > 0) {
+            levs_idx <- levs_idx[idx, , drop = FALSE]
+        }
     }
 
     levs_idx$isNULL <- levs_idx$hi == levs_idx$lo